Moved runtime generation back to JSExecutionInterface
