+++ src/main/java/com/violeth/blockjs/blockjs/jsinterface/JSExecutionInterface.java
@@ -1,5 +1,6 @@
 package com.violeth.blockjs.blockjs.jsinterface;
 
+import com.caoccao.javet.enums.JSRuntimeType;
 import com.caoccao.javet.exceptions.JavetCompilationException;
 import com.caoccao.javet.exceptions.JavetException;
 import com.caoccao.javet.exceptions.JavetExecutionException;
@@ -7,6 +8,8 @@ import com.caoccao.javet.interception.logging.JavetStandardConsoleInterceptor;
 import com.caoccao.javet.interop.NodeRuntime;
 import com.caoccao.javet.interop.V8Host;
 import com.caoccao.javet.interop.V8Runtime;
+import com.caoccao.javet.interop.engine.JavetEngineConfig;
+import com.caoccao.javet.interop.engine.JavetEnginePool;
 import com.violeth.blockjs.blockjs.BlockJS;
 import com.violeth.blockjs.blockjs.jsinterface.mcinterface.Chat;
 import com.violeth.blockjs.blockjs.jsinterface.mcinterface.Player;
@@ -16,19 +19,58 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.util.logging.Level;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 
 public final class JSExecutionInterface {
     public File file;
     public NodeRuntime runtime;
 
-    public JSExecutionInterface(File file, NodeRuntime runtime) {
+    public JSExecutionInterface(File file) {
         this.file = file;
-        this.runtime = runtime;
+    }
+
+    NodeRuntime createRuntime(File file) throws JavetException {
+        JavetEngineConfig config = new JavetEngineConfig();
+        config.setJSRuntimeType(JSRuntimeType.Node);
+
+        JavetEnginePool<NodeRuntime> pool = new JavetEnginePool<NodeRuntime>(config);
+
+        NodeRuntime runtime = pool.getEngine().getV8Runtime();
+
+        runtime.setV8ModuleResolver(((v8Runtime, resourceName, iv8Module) -> {
+            // Matches the file extension
+            Pattern pattern = Pattern.compile("((?<=\\w)\\.[^.]+)$");
+            Matcher matcher = pattern.matcher(resourceName);
+
+            String ext = matcher.find() ? matcher.group(1) : "";
+
+            String trueResourceName;
+
+            if (ext.equals("")) {
+                trueResourceName = resourceName + ".js";
+            } else {
+                trueResourceName = resourceName;
+            }
+
+            File resourceFile = new File(file.getParentFile(), trueResourceName);
+
+            if (!resourceFile.exists()) {
+                // TODO: Resolve module from node_modules
+                return null;
+            }
+
+            return runtime.getExecutor(resourceFile)
+                    .setResourceName(resourceName)
+                    .compileV8Module();
+        }));
+
+        return runtime;
     }
 
     public void registerAndRun() {
-        try {
+        try (NodeRuntime runtime = createRuntime(file)) {
             var consoleInterceptor = new JavetStandardConsoleInterceptor(runtime);
             consoleInterceptor.register(runtime.getGlobalObject());
