+++ src/main/java/com/violeth/blockjs/blockjs/jsinterface/JSRunner.java
@@ -23,44 +23,6 @@ public class JSRunner {
         this.folder = folder;
     }
 
-    NodeRuntime createRuntime(File file) throws JavetException {
-        JavetEngineConfig config = new JavetEngineConfig();
-        config.setJSRuntimeType(JSRuntimeType.Node);
-
-        JavetEnginePool<NodeRuntime> pool = new JavetEnginePool<NodeRuntime>(config);
-
-        NodeRuntime runtime = pool.getEngine().getV8Runtime();
-
-        runtime.setV8ModuleResolver(((v8Runtime, resourceName, iv8Module) -> {
-            // Matches the file extension
-            Pattern pattern = Pattern.compile("((?<=\\w)\\.[^.]+)$");
-            Matcher matcher = pattern.matcher(resourceName);
-
-            String ext = matcher.find() ? matcher.group(1) : "";
-
-            String trueResourceName;
-
-            if (ext.equals("")) {
-                trueResourceName = resourceName + ".js";
-            } else {
-                trueResourceName = resourceName;
-            }
-
-            File resourceFile = new File(file.getParentFile(), trueResourceName);
-
-            if (!resourceFile.exists()) {
-                // TODO: Resolve module from node_modules
-                return null;
-            }
-
-            return runtime.getExecutor(resourceFile)
-                    .setResourceName(resourceName)
-                    .compileV8Module();
-        }));
-
-        return runtime;
-    }
-
     public void getAndRunJS() {
         var files = FileReader.getListOfFiles(folder).toArray(new File[0]);
 
@@ -75,20 +37,16 @@ public class JSRunner {
         var interfaces = new JSExecutionInterface[files.length];
 
         for (var i = 0; i < files.length; i++) {
-            try (NodeRuntime runtime = createRuntime(files[i])) {
-                var file = files[i];
+            var file = files[i];
 
-                var ext = file.getName().substring(file.getName().lastIndexOf(".") + 1);
+            var ext = file.getName().substring(file.getName().lastIndexOf(".") + 1);
 
-                if (ext.equals("js")) {
-                    var jsInterface = new JSExecutionInterface(file.getAbsoluteFile(), runtime);
+            if (ext.equals("js")) {
+                var jsInterface = new JSExecutionInterface(file.getAbsoluteFile());
 
-                    interfaces[i] = jsInterface;
+                interfaces[i] = jsInterface;
 
-                    jsInterface.registerAndRun();
-                }
-            } catch (JavetException e) {
-                e.printStackTrace();
+                jsInterface.registerAndRun();
             }
         }
     }
