+++ src/test/java/com/violeth/blockjs/blockjs/jsinterface/JSExecutionInterfaceTest.java
@@ -30,11 +30,6 @@ class JSExecutionInterfaceTest {
         blockJS = MockBukkit.load(BlockJS.class);
     }
 
-    @BeforeEach
-    void setupEach() throws JavetException {
-        setupRuntime();
-    }
-
     @AfterAll
     static void teardown() {
         MockBukkit.unmock();
@@ -50,69 +45,24 @@ class JSExecutionInterfaceTest {
         return new File(Objects.requireNonNull(url).getFile());
     }
 
-    NodeRuntime setupRuntime() throws JavetException {
-        JavetEngineConfig config = new JavetEngineConfig();
-
-        config.setJSRuntimeType(JSRuntimeType.Node);
-
-        JavetEnginePool<NodeRuntime> pool = new JavetEnginePool<NodeRuntime>(config);
-
-        pool.getConfig().setJSRuntimeType(JSRuntimeType.Node);
-
-        NodeRuntime runtime = pool.getEngine().getV8Runtime();
-
-        return runtime;
-    }
-
     @Test
     @DisplayName("Can run JS")
     void runJS() {
-        try {
-            runner = new JSExecutionInterface(getResourceFile("jsTests/basicTest.js"), setupRuntime());
-            runner.runtime.await();
-            runner.registerAndRun();
-        } catch (JavetException e) {
-            BlockJS.instance.getLogger().log(Level.SEVERE,
-                    "There was an error creating the Javet runtime.");
-            throw new RuntimeException(e);
-        }
+        runner = new JSExecutionInterface(getResourceFile("jsTests/basicTest.js"));
+        runner.registerAndRun();
     }
 
     @Test
     @DisplayName("Can run JS with callbacks")
     void runJSWithCallbacks() {
-        try (NodeRuntime runtime = setupRuntime()) {
-            runner = new JSExecutionInterface(getResourceFile("jsTests/callbackTest.js"), runtime);
-            runner.runtime.await();
-            assertDoesNotThrow(() -> runner.registerAndRun());
-        } catch (JavetException e) {
-            BlockJS.instance.getLogger().log(Level.SEVERE,
-                    "There was an error creating the Javet runtime.");
-            throw new RuntimeException(e);
-        }
+        runner = new JSExecutionInterface(getResourceFile("jsTests/callbackTest.js"));
+        assertDoesNotThrow(() -> runner.registerAndRun());
     }
 
     @Test
     @DisplayName("Can run JS with imports")
     void runJSWithImports() {
-        try (NodeRuntime runtime = setupRuntime()) {
-            runtime.setV8ModuleResolver(((v8Runtime, resourceName, iv8Module) -> {
-                if (resourceName.equals("./importTestModule.js")) {
-                    return runtime.getExecutor(getResourceFile("jsTests/importTestModule.js"))
-                            .setResourceName(resourceName)
-                            .compileV8Module();
-                } else {
-                    return null;
-                }
-            }));
-
-            runner = new JSExecutionInterface(getResourceFile("jsTests/importTest.js"), runtime);
-            runner.runtime.await();
-            assertDoesNotThrow(() -> runner.registerAndRun());
-        } catch (JavetException e) {
-            BlockJS.instance.getLogger().log(Level.SEVERE,
-                    "There was an error creating the Javet runtime.");
-            throw new RuntimeException(e);
-        }
+        runner = new JSExecutionInterface(getResourceFile("jsTests/importTest.js"));
+        assertDoesNotThrow(() -> runner.registerAndRun());
     }
 }
